backend/requirements.txt
aioredis==2.0.1
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.18.0
elasticsearch==7.17.0
fastapi==0.109.2
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
identify==2.5.35
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
minio==7.2.5
motor==3.4.0
nodeenv==1.8.0
opencv-python==4.7.0.72
passlib==1.7.4
platformdirs==4.2.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pymongo==4.6.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.9.0
urllib3
uvicorn==0.27.0.post1
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
gunicorn==21.2.0
pandas==2.2.1
openpyxl==3.1.2
fastapi-pagination==0.12.24
bcrypt==4.1.2
Celery==5.4.0
pillow==10.3.0
jinja2==3.1.4
ruff==0.5.0
pre-commit==3.6.2
tritonclient[all]
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
sentry-sdk==2.18.0
aio-pika==9.4.3
pika==1.3.2
firebase-admin==6.6.0
nats-py==2.9.0
celery-batches==0.9
